*: Allow 16 bit size for nexthops
Currently FRR is limiting the nexthop count to a uint8_t not a
uint16_t.  This leads to issues when the nexthop count is 256
which results in the count to overflow to 0 causing problems
in the code.

+++ bgpd/bgp_attr.h
@@ -218,7 +218,7 @@ struct attr {
 	as_t aggregator_as;
 
 	/* MP Nexthop length */
-	uint8_t mp_nexthop_len;
+	uint16_t mp_nexthop_len;
 
 	/* MP Nexthop preference */
 	uint8_t mp_nexthop_prefer_global;
